lxc/project: add --force flag

+++ lxc/project.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"bufio"
 	"errors"
 	"fmt"
 	"io"
@@ -189,6 +190,9 @@ func (c *cmdProjectCreate) run(cmd *cobra.Command, args []string) error {
 type cmdProjectDelete struct {
 	global  *cmdGlobal
 	project *cmdProject
+
+	flagForce          bool
+	flagNonInteractive bool
 }
 
 func (c *cmdProjectDelete) command() *cobra.Command {
@@ -198,6 +202,9 @@ func (c *cmdProjectDelete) command() *cobra.Command {
 	cmd.Short = i18n.G("Delete projects")
 	cmd.Long = cli.FormatSection(i18n.G("Description"), i18n.G(
 		`Delete projects`))
+
+	cmd.Flags().BoolVarP(&c.flagForce, "force", "f", false, i18n.G("Force delete project and its entities (instances, profiles, images, networks, network ACLs, network zones, and storage volumes)")+"``")
+	cmd.Flags().BoolVar(&c.flagNonInteractive, "yes", false, i18n.G("Don't require user confirmation for using --force"))
 
 	cmd.RunE = c.run
 
@@ -210,6 +217,21 @@ func (c *cmdProjectDelete) command() *cobra.Command {
 	}
 
 	return cmd
+}
+
+func (c *cmdProjectDelete) promptConfirmation(name string) error {
+	reader := bufio.NewReader(os.Stdin)
+	fmt.Printf(i18n.G(`Forcefully removing a project will delete the project and its entities (instances, profiles, images, networks, network ACLs, network zones, and storage volumes)
+
+Are you really sure you want to force removing %s? (yes/no): `), name)
+	input, _ := reader.ReadString('\n')
+	input = strings.TrimSuffix(input, "\n")
+
+	if !slices.Contains([]string{i18n.G("yes")}, strings.ToLower(input)) {
+		return errors.New(i18n.G("User aborted delete operation"))
+	}
+
+	return nil
 }
 
 func (c *cmdProjectDelete) run(cmd *cobra.Command, args []string) error {
@@ -236,10 +258,25 @@ func (c *cmdProjectDelete) run(cmd *cobra.Command, args []string) error {
 		return errors.New(i18n.G("Missing project name"))
 	}
 
-	// Delete the project
-	err = resource.server.DeleteProject(resource.name)
-	if err != nil {
-		return err
+	// Delete the project.
+	if c.flagForce {
+		if !c.flagNonInteractive {
+			// Prompt for confirmation if --force is used.
+			err := c.promptConfirmation(resource.name)
+			if err != nil {
+				return err
+			}
+		}
+
+		err := resource.server.ForceDeleteProject(resource.name)
+		if err != nil {
+			return err
+		}
+	} else {
+		err = resource.server.DeleteProject(resource.name)
+		if err != nil {
+			return err
+		}
 	}
 
 	if !c.global.flagQuiet {
